im probably never touching this again as its too much work, i will just do all this in self-hosted rewrite for now it doesnt matter if the code is s**t

+++ steelc/src/lexer/token_utils.cpp
@@ -148,7 +148,7 @@ int op_len(token_type tk) {
 	const auto& operators = get_operators();
 	for (const auto& op : operators) {
 		if (op.second == tk) {
-			return op.first.size();
+			return static_cast<int>(op.first.size());
 		}
 	}
 	return 0;
